+++ src/test/java/com/github/jnidzwetzki/bitfinex/v2/test/WalletHandlerTest.java
@@ -54,7 +54,7 @@ public class WalletHandlerTest {
 		final Table<String, String, Wallet> walletTable = HashBasedTable.create();
 
 		final BitfinexApiBroker bitfinexApiBroker = Mockito.mock(BitfinexApiBroker.class);
-		Mockito.when(bitfinexApiBroker.getWalletTable()).thenReturn(walletTable);
+		Mockito.when(bitfinexApiBroker.getWalletManager().getWalletTable()).thenReturn(walletTable);
 		Mockito.when(bitfinexApiBroker.getConnectionReadyLatch()).thenReturn(walletLatch);
 		
 		Assert.assertTrue(walletTable.isEmpty());
@@ -84,7 +84,7 @@ public class WalletHandlerTest {
 		final Table<String, String, Wallet> walletTable = HashBasedTable.create();
 
 		final BitfinexApiBroker bitfinexApiBroker = Mockito.mock(BitfinexApiBroker.class);
-		Mockito.when(bitfinexApiBroker.getWalletTable()).thenReturn(walletTable);
+		Mockito.when(bitfinexApiBroker.getWalletManager().getWalletTable()).thenReturn(walletTable);
 		Mockito.when(bitfinexApiBroker.getConnectionReadyLatch()).thenReturn(walletLatch);
 		
 		Assert.assertTrue(walletTable.isEmpty());