+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 0.6.1 (TBA)
+* New Feature: Moved the wallets into a WalletManager
 * Improvement: Switched to tyrus-standalone-client-jdk (uses Java SE 7 Asynchronous IO) for increased performance
 * Bugfix: Removed invalid currency symbol (closes #18)